Added variable "identifier" to --each-template
In some cases its necessary to adjust the x- and y-values for the background-positions of some sprites. Of example to center a sprite image which is used for bullet points.

With the variable "identifier" its possible to create @LessVariables for x- and y-values in each-template to use them in own Less files.